refactor: relocate `get_parameters` function to utilitys so it can be used by other classes!
